Get compilation with more exhaustiveness working again,
but this time with a cleaner story, since in PhiCt we already only have
ClassIds.

It works for the first two examples in #19272, but it does not solve the
latest contrived example we came up with -- I think it's likely because
of rebuilding.

RE rebuilding: It doesn't look like usually there are many/any things in
the worklist, meaning it's probably quite cheap to rebuild often.