Make list comprehension completely non-linear
Fixes #25081

From the note:

The usefulness of list comprehension in conjunction with linear types is dubious.
After all, statements are made to be run many times, for instance in

```haskell
[u | y <- [0,1], stmts]
```

both `u` and `stmts` are going to be run several times.

In principle, though, there are some position in a monad comprehension
expression which could be considered linear. We could try and make it so that
these positions are considered linear by the typechecker, but in practice the
desugarer doesn't take enough care to ensure that these are indeed desugared to
linear sites. We tried in the past, and it turned out that we'd miss a
desugaring corner case (#25772).

Until there's a demand for this very specific improvement, let's instead be
conservative, and consider list comprehension to be completely non-linear.