Update and generate libHSghc-internal.def from .def.in file
The only symbol that the rts imports from the ghc-internal package now
is init_ghc_hs_iface. So the rts only needs an import lib that defines
that one symbol.

Also, remove the libHSghc-prim.def because it is redundant. The rts no
longer imports anything from ghc-prim.

Keep libHSffi.def for now. We may yet need it once it is clear how
libffi is going to be built/used for ghc.